Automatically "bake" transforms
When applying a transform via the rotate, translate, scale added to the
Element prototype, the returned object is a different one, with the
points in the location they would be after the transformation has been
applied (or "baked" into the element).
Alternatively the Element.flatten() function can be called to flatten
any transformations applied to the element in other ways.